Compacted ProjectGrid and ProjectCard into one file
