+++ src/components/Projects/index.js
@@ -1,13 +1,43 @@
 import React from "react"
+import Img from "gatsby-image"
+import { Link } from "gatsby"
+import NewPost from "../NewPost"
 import { StaticQuery, graphql } from "gatsby"
-import ProjectCard from "../ProjectCard/index"
 
-import "./ProjectGrid.scss"
+import "./Card.scss"
+import "./Grid.scss"
 
-const ProjectGrid = () => (
+export const Card = ({ client, slug, img, year, categories, id }) => {
+  const postYear = year.slice(3, 7)
+
+  return (
+    <div className="project-card" data-sal="fade" key={id}>
+      <Link to={slug}>
+        <NewPost />
+        <Img
+          className="project--featured-media"
+          fluid={img}
+          loading="lazy"
+          alt={client}
+          objectPosition="50% 50%"
+          placeholderStyle={{ backgroundColor: "#242424" }}
+        />
+        <div className="project--information">
+          <h3>{client}</h3>
+          <div className="right">
+            <p>{postYear}</p>
+            <p>{categories}</p>
+          </div>
+        </div>
+      </Link>
+    </div>
+  )
+}
+
+export const Grid = () => (
   <StaticQuery
     query={graphql`
-      query PROJECTGRID_QUERY {
+      query PROJECTGRID_QUERY_NEW {
         allWordpressPost {
           edges {
             node {
@@ -36,7 +66,7 @@ const ProjectGrid = () => (
       <section className="project-grid">
         <h2>Prosjekter</h2>
         {data.allWordpressPost.edges.map((item) => (
-          <ProjectCard
+          <Card
             id={item.node.id}
             slug={`/${item.node.slug}`}
             client={item.node.acf.client}
@@ -49,5 +79,3 @@ const ProjectGrid = () => (
     )}
   />
 )
-
-export default ProjectGrid
